homework5.1-5.2

+++ link-parser/src/main/java/ru/tinkoff/edu/java/linkparser/dto/StackOverflowDTO.java
@@ -1,4 +1,4 @@
 package ru.tinkoff.edu.java.linkparser.dto;
 
-public record StackOverflowDTO(int questionId) implements ILinkDTO {
+public record StackOverflowDTO(long questionId) implements ILinkDTO {
 }
