+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/client/StackOverflowClient.java
@@ -4,5 +4,7 @@ import reactor.core.publisher.Mono;
 import ru.tinkoff.edu.java.scrapper.dto.response.StackOverflowResponse;
 
 public interface StackOverflowClient {
-    Mono<StackOverflowResponse> fetchQuestion(int questionId);
+
+    Mono<StackOverflowResponse> fetchQuestion(long questionId);
+
 }
